+++ src/Amazon.SQS.ExtendedClient/Properties.cs
@@ -1,0 +1,3 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Amazon.SQS.ExtendedClient.Tests")]
