--- src/Amazon.SQS.ExtendedClient/Properties/AssemblyInfo.cs
@@ -1,37 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("Amazon.SQS.ExtendedClient")]
-[assembly: AssemblyDescription("Extension to Amazon SQS that adds support for sending and receiving messages greater than 256K")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Oleg Rakitskiy")]
-[assembly: AssemblyProduct("Amazon.SQS.ExtendedClient")]
-[assembly: AssemblyCopyright("Copyright © Oleg Rakitskiy 2016")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-// Setting ComVisible to false makes the types in this assembly not visible 
-// to COM components.  If you need to access a type in this assembly from 
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("38832a23-5aea-4856-991e-81ed4082bb67")]
-
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version 
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers 
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("0.1.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
-[assembly: InternalsVisibleTo("Amazon.SQS.ExtendedClient.Tests")]
