+++ test/Amazon.SQS.ExtendedClient.Tests/Amazon.SQS.ExtendedClient.Tests.csproj
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>netcoreapp1.1</TargetFramework>
+    <DebugType>portable</DebugType>
+    <AssemblyName>Amazon.SQS.ExtendedClient.Tests</AssemblyName>
+    <PackageId>Amazon.SQS.ExtendedClient.Tests</PackageId>
+    <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
+    <PackageTargetFallback>$(PackageTargetFallback);portable-net451+win8</PackageTargetFallback>
+    <RuntimeFrameworkVersion>1.1.1</RuntimeFrameworkVersion>
+  </PropertyGroup>
+  <ItemGroup>
+    <ProjectReference Include="../../src/Amazon.SQS.ExtendedClient/Amazon.SQS.ExtendedClient.csproj"/>
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.0.0"/>
+    <PackageReference Include="NUnit" Version="3.5.0"/>
+    <PackageReference Include="dotnet-test-nunit" Version="3.4.0-beta-3"/>
+    <PackageReference Include="NUnit3TestAdapter" Version="3.8.0-alpha1"/>
+    <PackageReference Include="Moq" Version="4.6.38-alpha"/>
+    <PackageReference Include="Microsoft.DotNet.InternalAbstractions" Version="1.0.500-preview2-1-003177"/>
+    <PackageReference Include="AWSSDK.Core" Version="3.3.17.10"/>
+    <PackageReference Include="AWSSDK.S3" Version="3.3.11"/>
+    <PackageReference Include="AWSSDK.SQS" Version="3.3.2.7"/>
+    <PackageReference Include="Newtonsoft.Json" Version="10.0.3"/>
+  </ItemGroup>
+</Project>
