+++ test/Amazon.SQS.ExtendedClient.Tests/Constants.cs
