+++ test/Amazon.SQS.ExtendedClient.Tests/packages.config
